More changes for real GZ server to work

+++ db/migrate/20140915081141_set_first_user_quick_performers.rb
@@ -1,11 +1,13 @@
 class SetFirstUserQuickPerformers < ActiveRecord::Migration
   def change
     u = User.first
-    u.quick_performers.destroy_all
-    index = 0
-    TakeOffice::Employee.quick_performers.each { |employee| 
-      u.quick_performers << QuickPerformer.create!(employee: employee, order: index)
-      index += 1
-    }
+    unless u.nil?
+      u.quick_performers.destroy_all
+      index = 0
+      TakeOffice::Employee.quick_performers.each { |employee| 
+        u.quick_performers << QuickPerformer.create!(employee: employee, order: index)
+        index += 1
+      }
+    end
   end
 end
